add safe stream config
