+++ aether/stream_api/safe_stream/safe_stream_receiving.h
@@ -29,6 +29,7 @@
 #include "aether/transport/data_buffer.h"
 #include "aether/stream_api/safe_stream/safe_stream_api.h"
 #include "aether/stream_api/safe_stream/safe_stream_types.h"
+#include "aether/stream_api/safe_stream/safe_stream_config.h"
 
 namespace ae {
